got mock testing working
