+++ package.json
@@ -4,8 +4,8 @@
   "description": "a module for integrating amazon dash button presses into a node project",
   "main": "index.js",
   "scripts": {
-    "livetest": "mocha test/test.js",
-    "test": "mocha test/test-mocks.js"
+    "livetest": "istanbul cover ./node_modules/mocha/bin/_mocha test/test-live.js --report lcovonly -- -R spec && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage",
+    "test": "istanbul cover ./node_modules/mocha/bin/_mocha test/test.js --report lcovonly -- -R spec && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage"
   },
   "bin":{
     "findbutton": "./bin/findbutton"
@@ -39,7 +39,9 @@
     "underscore": "^1.8.3",
     "sinon": "*",
     "mockery": "*",
-    "ip":"*"
+    "ip":"*",
+    "istanbul": "*",
+    "coveralls": "*"
   }
 
 }
